NNUE Architecture

+++ src/search.cpp
@@ -14,7 +14,7 @@ void SearchThread::order_moves(MoveList &moves, int nr_moves)
     {
         Move move = moves[i];
         if (board.is_capture(move))
-            scores[i] = 1000 * board.at(move.to());
+            scores[i] = 1000 * (uint8_t)board.at(move.to());
         else
             scores[i] = 0;
     }
